Convert table view cells to collection view cells
