+++ kDrive/UI/View/Menu/Store/StoreControlCollectionViewCell.xib
@@ -1,0 +1,54 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="19529" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
+    <device id="retina6_1" orientation="portrait" appearance="light"/>
+    <dependencies>
+        <deployment identifier="iOS"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="19519"/>
+        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
+        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
+    </dependencies>
+    <objects>
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
+        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
+        <collectionViewCell opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" id="gTV-IL-0wX" customClass="StoreControlCollectionViewCell" customModule="kDrive" customModuleProvider="target">
+            <rect key="frame" x="0.0" y="0.0" width="320" height="40"/>
+            <autoresizingMask key="autoresizingMask"/>
+            <view key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center">
+                <rect key="frame" x="0.0" y="0.0" width="320" height="40"/>
+                <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
+                <subviews>
+                    <segmentedControl opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="left" contentVerticalAlignment="top" segmentControlStyle="plain" selectedSegmentIndex="0" translatesAutoresizingMaskIntoConstraints="NO" id="4pq-It-6k7" customClass="IKSegmentedControl" customModule="kDriveCore">
+                        <rect key="frame" x="24" y="0.0" width="272" height="41"/>
+                        <constraints>
+                            <constraint firstAttribute="height" constant="40" id="3ja-t3-D26"/>
+                        </constraints>
+                        <segments>
+                            <segment title="First"/>
+                            <segment title="Second"/>
+                        </segments>
+                        <connections>
+                            <action selector="segmentedControlValueChanged:" destination="gTV-IL-0wX" eventType="valueChanged" id="QZb-Qz-rps"/>
+                        </connections>
+                    </segmentedControl>
+                </subviews>
+            </view>
+            <viewLayoutGuide key="safeArea" id="ZTg-uK-7eu"/>
+            <constraints>
+                <constraint firstAttribute="bottom" secondItem="4pq-It-6k7" secondAttribute="bottom" id="7Nt-Gm-uwh"/>
+                <constraint firstItem="4pq-It-6k7" firstAttribute="top" secondItem="gTV-IL-0wX" secondAttribute="top" id="GB8-nk-Pdb"/>
+                <constraint firstAttribute="trailing" secondItem="4pq-It-6k7" secondAttribute="trailing" constant="24" id="GzC-9d-gwz"/>
+                <constraint firstItem="4pq-It-6k7" firstAttribute="leading" secondItem="gTV-IL-0wX" secondAttribute="leading" constant="24" id="ons-3W-zyZ"/>
+            </constraints>
+            <size key="customSize" width="232" height="26"/>
+            <connections>
+                <outlet property="segmentedControl" destination="4pq-It-6k7" id="4IY-JA-he4"/>
+            </connections>
+            <point key="canvasLocation" x="263.768115942029" y="70.3125"/>
+        </collectionViewCell>
+    </objects>
+    <designables>
+        <designable name="4pq-It-6k7">
+            <size key="intrinsicContentSize" width="131" height="31"/>
+        </designable>
+    </designables>
+</document>
