--- kDrive/UI/View/Menu/Store/StoreControlTableViewCell.xib
@@ -1,54 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="18122" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
-    <device id="retina6_1" orientation="portrait" appearance="light"/>
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="18093"/>
-        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
-        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <tableViewCell contentMode="scaleToFill" selectionStyle="default" indentationWidth="10" rowHeight="63" id="KGk-i7-Jjw" customClass="StoreControlTableViewCell" customModule="kDrive" customModuleProvider="target">
-            <rect key="frame" x="0.0" y="0.0" width="320" height="64"/>
-            <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
-            <tableViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" tableViewCell="KGk-i7-Jjw" id="H2p-sc-9uM">
-                <rect key="frame" x="0.0" y="0.0" width="320" height="64"/>
-                <autoresizingMask key="autoresizingMask"/>
-                <subviews>
-                    <segmentedControl opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="left" contentVerticalAlignment="top" segmentControlStyle="plain" selectedSegmentIndex="0" translatesAutoresizingMaskIntoConstraints="NO" id="fBY-KL-Qvd" customClass="IKSegmentedControl" customModule="kDriveCore">
-                        <rect key="frame" x="24" y="0.0" width="272" height="41"/>
-                        <constraints>
-                            <constraint firstAttribute="height" constant="40" id="tX5-mM-xKY"/>
-                        </constraints>
-                        <segments>
-                            <segment title="First"/>
-                            <segment title="Second"/>
-                        </segments>
-                        <connections>
-                            <action selector="segmentedControlValueChanged:" destination="KGk-i7-Jjw" eventType="valueChanged" id="xT1-6Z-Urg"/>
-                        </connections>
-                    </segmentedControl>
-                </subviews>
-                <constraints>
-                    <constraint firstAttribute="trailing" secondItem="fBY-KL-Qvd" secondAttribute="trailing" constant="24" id="5rl-m3-gsa"/>
-                    <constraint firstItem="fBY-KL-Qvd" firstAttribute="leading" secondItem="H2p-sc-9uM" secondAttribute="leading" constant="24" id="nSD-SN-tJ4"/>
-                    <constraint firstItem="fBY-KL-Qvd" firstAttribute="top" secondItem="H2p-sc-9uM" secondAttribute="top" id="w8g-MW-kx7"/>
-                    <constraint firstAttribute="bottom" secondItem="fBY-KL-Qvd" secondAttribute="bottom" constant="24" id="xu4-h1-e8t"/>
-                </constraints>
-            </tableViewCellContentView>
-            <viewLayoutGuide key="safeArea" id="njF-e1-oar"/>
-            <color key="backgroundColor" white="0.0" alpha="0.0" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-            <connections>
-                <outlet property="segmentedControl" destination="fBY-KL-Qvd" id="oSI-ai-MKY"/>
-            </connections>
-            <point key="canvasLocation" x="131.8840579710145" y="79.352678571428569"/>
-        </tableViewCell>
-    </objects>
-    <designables>
-        <designable name="fBY-KL-Qvd">
-            <size key="intrinsicContentSize" width="131" height="31"/>
-        </designable>
-    </designables>
-</document>
