+++ kDrive/UI/View/Menu/Store/StoreNextCollectionViewCell.xib
@@ -1,0 +1,54 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="19529" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
+    <device id="retina6_1" orientation="portrait" appearance="light"/>
+    <dependencies>
+        <deployment identifier="iOS"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="19519"/>
+        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
+        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
+    </dependencies>
+    <objects>
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
+        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
+        <collectionViewCell opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" id="gTV-IL-0wX" customClass="StoreNextCollectionViewCell" customModule="kDrive" customModuleProvider="target">
+            <rect key="frame" x="0.0" y="0.0" width="320" height="68"/>
+            <autoresizingMask key="autoresizingMask"/>
+            <view key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center">
+                <rect key="frame" x="0.0" y="0.0" width="320" height="68"/>
+                <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
+                <subviews>
+                    <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="system" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="FRY-jy-Ld4" customClass="IKLargeButton" customModule="kDriveCore">
+                        <rect key="frame" x="24" y="8" width="272" height="60"/>
+                        <constraints>
+                            <constraint firstAttribute="height" constant="60" id="bVN-3E-VuY"/>
+                        </constraints>
+                        <state key="normal" title="Suivant"/>
+                        <userDefinedRuntimeAttributes>
+                            <userDefinedRuntimeAttribute type="string" keyPath="xibLocKey" value="buttonNext"/>
+                        </userDefinedRuntimeAttributes>
+                        <connections>
+                            <action selector="buttonTapped:" destination="gTV-IL-0wX" eventType="touchUpInside" id="hG9-Y3-N1q"/>
+                        </connections>
+                    </button>
+                </subviews>
+            </view>
+            <viewLayoutGuide key="safeArea" id="ZTg-uK-7eu"/>
+            <constraints>
+                <constraint firstItem="FRY-jy-Ld4" firstAttribute="top" secondItem="gTV-IL-0wX" secondAttribute="top" constant="8" id="8CD-aH-sLu"/>
+                <constraint firstAttribute="bottom" secondItem="FRY-jy-Ld4" secondAttribute="bottom" id="Ko0-UV-RW1"/>
+                <constraint firstAttribute="trailing" secondItem="FRY-jy-Ld4" secondAttribute="trailing" constant="24" id="Tnn-bQ-z3o"/>
+                <constraint firstItem="FRY-jy-Ld4" firstAttribute="leading" secondItem="gTV-IL-0wX" secondAttribute="leading" constant="24" id="v0G-Zu-fXd"/>
+            </constraints>
+            <size key="customSize" width="320" height="68"/>
+            <connections>
+                <outlet property="button" destination="FRY-jy-Ld4" id="ahh-Ud-HiU"/>
+            </connections>
+            <point key="canvasLocation" x="131.8840579710145" y="82.366071428571431"/>
+        </collectionViewCell>
+    </objects>
+    <designables>
+        <designable name="FRY-jy-Ld4">
+            <size key="intrinsicContentSize" width="75" height="29"/>
+        </designable>
+    </designables>
+</document>
