--- kDrive/UI/View/Menu/Store/StoreOffersTableViewCell.xib
@@ -1,54 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="18122" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES">
-    <device id="retina6_1" orientation="portrait" appearance="light"/>
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="18093"/>
-        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <objects>
-        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner"/>
-        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
-        <tableViewCell contentMode="scaleToFill" selectionStyle="default" indentationWidth="10" id="KGk-i7-Jjw" customClass="StoreOffersTableViewCell" customModule="kDrive" customModuleProvider="target">
-            <rect key="frame" x="0.0" y="0.0" width="320" height="360"/>
-            <autoresizingMask key="autoresizingMask" flexibleMaxX="YES" flexibleMaxY="YES"/>
-            <tableViewCellContentView key="contentView" opaque="NO" clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="center" tableViewCell="KGk-i7-Jjw" id="H2p-sc-9uM">
-                <rect key="frame" x="0.0" y="0.0" width="320" height="360"/>
-                <autoresizingMask key="autoresizingMask"/>
-                <subviews>
-                    <collectionView clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" alwaysBounceHorizontal="YES" showsHorizontalScrollIndicator="NO" showsVerticalScrollIndicator="NO" dataMode="none" translatesAutoresizingMaskIntoConstraints="NO" id="1fk-pm-GjC">
-                        <rect key="frame" x="0.0" y="0.0" width="320" height="360"/>
-                        <color key="backgroundColor" white="0.0" alpha="0.0" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-                        <constraints>
-                            <constraint firstAttribute="height" constant="360" id="QFS-bd-HKv"/>
-                        </constraints>
-                        <collectionViewFlowLayout key="collectionViewLayout" scrollDirection="horizontal" minimumLineSpacing="10" minimumInteritemSpacing="10" id="IUN-d9-daz">
-                            <size key="itemSize" width="128" height="128"/>
-                            <size key="headerReferenceSize" width="0.0" height="0.0"/>
-                            <size key="footerReferenceSize" width="0.0" height="0.0"/>
-                            <inset key="sectionInset" minX="24" minY="0.0" maxX="24" maxY="0.0"/>
-                        </collectionViewFlowLayout>
-                        <cells/>
-                        <connections>
-                            <outlet property="dataSource" destination="KGk-i7-Jjw" id="uF8-0S-irr"/>
-                            <outlet property="delegate" destination="KGk-i7-Jjw" id="AfR-Yu-ypn"/>
-                        </connections>
-                    </collectionView>
-                </subviews>
-                <constraints>
-                    <constraint firstItem="1fk-pm-GjC" firstAttribute="leading" secondItem="H2p-sc-9uM" secondAttribute="leading" id="9fx-6h-AQx"/>
-                    <constraint firstAttribute="trailing" secondItem="1fk-pm-GjC" secondAttribute="trailing" id="QgM-3X-maq"/>
-                    <constraint firstItem="1fk-pm-GjC" firstAttribute="top" secondItem="H2p-sc-9uM" secondAttribute="top" id="gRj-WI-UpV"/>
-                    <constraint firstAttribute="bottom" secondItem="1fk-pm-GjC" secondAttribute="bottom" id="hpa-mJ-0op"/>
-                </constraints>
-            </tableViewCellContentView>
-            <viewLayoutGuide key="safeArea" id="njF-e1-oar"/>
-            <color key="backgroundColor" white="0.0" alpha="0.0" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
-            <connections>
-                <outlet property="collectionView" destination="1fk-pm-GjC" id="1mB-eR-BzY"/>
-            </connections>
-            <point key="canvasLocation" x="132" y="78"/>
-        </tableViewCell>
-    </objects>
-</document>
